Rendering back to working as previously
