+++ Game/game.html
@@ -4,6 +4,6 @@
 </head>
  <body>
 	 <canvas id="game-canvas"></canvas>
-	 <script type="module" src="render.js"></script> 
+	 <script type="module" src="game.js"></script> 
  </body>
 </html>
